viewUriRoot - override where swagger-ui shows in application
When viewUriRoot is configured, the swagger-ui and assets are moved from
Jersey Context Root to the viewUriRoot. This allows Jersey Context to
be one value, uriPrefix (used for generating REST calls to our api) to
be something else, and the view to be somewhere else.

messy, but seems to work for viewUriRoot case

added docs for viewUriRoot